Fix object map section parsing for AC1032 format
Sample AC1032 fomat files are a bit peculiar - Handles section size is larger
than actual contents (looks like it is filled with padding). In the past we
would base on section size, but now we are a bit smarter and stop reading
object map sections upon finding empty section (as per ODA spec).