feat: expose more functions for class DRW_Header and use 'return_value_policy::reference' for class properties to avoid memory leak (#8)
By default property() bindings to objects use return_value_policy::copy which can very easily lead to memory leaks since each access to the property will create a new object that must be deleted. Alternatively, use return_value_policy::reference, so a new object is not allocated and changes to the object will be reflected in the original object.

https://emscripten.org/docs/porting/connecting_cpp_and_javascript/embind.html#class-properties